refactor: rename aibudget package to budget
